+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Not yet released
+
+- fix: Map columns correctly when exporting database
+
 ## v1.1.0 (09/12/2024)
 
 - fix: Fix check for existing db file